requirements.txt
camelot-py
opencv-python
pandas
pikepdf
beautifultable==1.0.1